before *Searching Articles*

+++ mysrio_blog/article/views.py
@@ -4,15 +4,29 @@ from .models import ArticlePost
 from .forms import ArticlePostForm
 from django.contrib.auth.models import User
 import markdown
+from django.contrib.auth.decorators import login_required
+from django.core.paginator import Paginator
 
 
 def article_list(request):
-    articles = ArticlePost.objects.all()  # ArticlePost.objects.all()是数据类的方法，可以获得所有的对象（即博客文章）
-    context = {'articles': articles}
-    # return HttpResponse("Response content: Hello Ferax!")
+    # 排序，需要GET请求中就已经给出order。GET请求也是可以传递多个参数的，如 ?a=1&b=2，参数间用&隔开
+    if request.GET.get('order') == 'total_views':
+        articles = ArticlePost.objects.all().order_by('-total_views')
+        # ‘total_views’为正序，‘-total_views’为逆序
+        order = 'total_views'
+    else:
+        articles = ArticlePost.objects.all()
+        order = 'normal'
 
-    # 最后返回了render函数。它的作用是结合模板和上下文，并返回渲染后的HttpResponse对象。
-    # 通俗的讲就是把context的内容，加载进模板，并通过浏览器呈现
+    # 每页显示 4 篇文章
+    paginator = Paginator(articles, 4)
+
+    # 获取 url 中的页码 (?key=value), 将导航对象相应的页码内容返回给 articles
+    page = request.GET.get('page')
+    articles = paginator.get_page(page)
+
+    context = {'articles': articles, 'order': order}
+    # 为什么把新变量order也传递到模板中？因为文章需要翻页！order给模板一个标识，提醒模板下一页应该如何排序
     return render(request, 'article/list.html', context)
 
 
@@ -25,14 +39,16 @@ def article_detail(request, id):
             "markdown.extensions.codehilite"
         ]
     )
+    article.total_views += 1
+    article.save(update_fields=['total_views'])  # update_fields=[]指定了数据库只更新total_views字段，优化执行效率。
     context = {'article': article}
     return render(request, 'article/detail.html', context)
 
 
+@login_required(login_url='/userprofile/login/')
 def article_create(request):
     """
     created和updated字段为自动生成，不需要填入；
-    author字段暂时固定为id=1的管理员用户，也不用填入；
     剩下的title和body就是*表单*需要填入的内容了
     :param request:
     :return:
@@ -47,7 +63,7 @@ def article_create(request):
         if article_post_form.is_valid():
             # 暂时还不提交数据库
             new_article = article_post_form.save(commit=False)
-            new_article.author = User.objects.get(id=1)
+            new_article.author = User.objects.get(id=request.user.id)
             new_article.save()  # 保存到数据库
             return redirect("article:article_list")  # redirect可通过url地址的名字，反向解析到对应的url。
         else:
@@ -59,17 +75,19 @@ def article_create(request):
         return render(request, 'article/create.html', context)
 
 
-
+@login_required(login_url='/userprofile/login/')
 def article_safe_delete(request, id):
     if request.method == 'POST':
         article = ArticlePost.objects.get(id=id)
+        if article.author.id != request.user.id:
+            return HttpResponse("不能删除其他人的文章")
         article.delete()
         return redirect("article:article_list")
     else:
         return HttpResponse("仅允许post请求")
 
 
-
+@login_required(login_url='/userprofile/login/')
 def article_update(request, id):
     """
     更新文章的视图函数
@@ -77,6 +95,9 @@ def article_update(request, id):
     id: article id
     """
     article = ArticlePost.objects.get(id=id)
+    if article.author.id != request.user.id:
+        return HttpResponse("不能修改其他人的文章")
+
     if request.method == "POST":
         article_post_form = ArticlePostForm(data=request.POST)
         if article_post_form.is_valid():
